Completed fron, employees and store not showing

+++ actions/locations/delete.ts
@@ -1,15 +1,14 @@
-"user server";
+"use server";
 
-import { API_URL} from "@/constants";
-import { AuthHeaders } from "@/helpers/authHelper";
-import { revalidatePath, revalidateTag } from "next/cache";
+import { API_URL } from "@/constants";
+import { AuthHeaders } from "@/helpers/authHeaders";
+import { revalidateTag } from "next/cache";
 import { redirect } from "next/navigation";
 
-export default async function DeleteLocation(formData: FormData){
-
-    const storeToDelete = formData.get("deleteValue"); 
-    fetch(`${API_URL}/locations/${storeToDelete}`,{
-
+export default async function deleteLocation(formData: FormData){
+  const locationId = formData.get("deleteValue")
+  if (!locationId) return;
+  fetch(`${API_URL}/locations/${locationId}`, {
         method:"DELETE",
         headers:{
              ... AuthHeaders(),
